+++ temba/campaigns/tests.py
@@ -712,18 +712,6 @@ class CampaignTest(TembaTest):
         self.set_contact_field(self.nonfarmer, "planting_date", "1/7/2025", legacy_handle=True)
         self.assertEqual(2, EventFire.objects.filter(event__is_active=True).count())
 
-        # remove one of the farmers from the group
-        self.farmers.update_contacts(self.admin, [self.farmer1], add=False)
-
-        # should only be one event now (on farmer 2)
-        fire = EventFire.objects.get()
-        self.assertEqual(2, fire.scheduled.day)
-        self.assertEqual(6, fire.scheduled.month)
-        self.assertEqual(2022, fire.scheduled.year)
-
-        # but if we add him back in, should be updated
-        post_data = dict(name=self.farmer1.name, groups=[self.farmers.id], __urn__tel=self.farmer1.get_urn("tel").path)
-
         planting_date_field = ContactField.get_by_key(self.org, "planting_date")
 
         self.client.post(reverse("contacts.contact_update", args=[self.farmer1.id]), post_data)
@@ -804,7 +792,7 @@ class CampaignTest(TembaTest):
 
         # page title and main content title should NOT contain (Archived)
         self.assertContains(response, "Perform the rain dance", count=2)
-        self.assertContains(response, "(Archived)", count=0)
+        self.assertContains(response, "Archived", count=0)
 
         gear_links = response.context["view"].get_gear_links()
         self.assertListEqual([gl["title"] for gl in gear_links], ["Add Event", "Export", "Edit", "Archive"])
@@ -817,7 +805,7 @@ class CampaignTest(TembaTest):
 
         # page title and main content title should contain (Archived)
         self.assertContains(response, "Perform the rain dance", count=2)
-        self.assertContains(response, "(Archived)", count=2)
+        self.assertContains(response, "Archived", count=2)
 
         gear_links = response.context["view"].get_gear_links()
         self.assertListEqual([gl["title"] for gl in gear_links], ["Activate", "Export"])
@@ -866,9 +854,9 @@ class CampaignTest(TembaTest):
 
         response = self.client.get(reverse("campaigns.campaignevent_read", args=[event.pk]))
 
-        # page title and main content title should NOT contain (Archived)
-        self.assertContains(response, "Perform the rain dance", count=2)
-        self.assertContains(response, "(Archived)", count=0)
+        # page title and main content title should NOT contain Archived
+        self.assertContains(response, "Perform the rain dance", count=1)
+        self.assertContains(response, "Archived", count=0)
 
         gear_links = response.context["view"].get_gear_links()
         self.assertListEqual([gl["title"] for gl in gear_links], ["Edit", "Delete"])
@@ -879,9 +867,9 @@ class CampaignTest(TembaTest):
 
         response = self.client.get(reverse("campaigns.campaignevent_read", args=[event.pk]))
 
-        # page title and main content title should contain (Archived)
-        self.assertContains(response, "Perform the rain dance", count=2)
-        self.assertContains(response, "(Archived)", count=1)
+        # page title and main content title should contain Archived
+        self.assertContains(response, "Perform the rain dance", count=1)
+        self.assertContains(response, "Archived", count=1)
 
         gear_links = response.context["view"].get_gear_links()
         self.assertListEqual([gl["title"] for gl in gear_links], ["Delete"])