+++ temba/contacts/tests.py
@@ -436,14 +436,10 @@ class ContactGroupTest(TembaTest):
         with MockParseQuery("id = 123", ["id"], allow_as_group=False):
             self.assertRaises(ValueError, ContactGroup.create_dynamic, self.org, self.admin, "Bose", "id = 123")
 
-        # can't call update_contacts on a dynamic group
-        self.assertRaises(ValueError, group.update_contacts, self.admin, [self.joe], True)
-
         # dynamic group should not have remove to group button
         self.login(self.admin)
         filter_url = reverse("contacts.contact_filter", args=[group.uuid])
         response = self.client.get(filter_url)
-        self.assertNotIn("unlabel", response.context["actions"])
         self.assertEqual(list(response.context["contact_fields"].values_list("key", flat=True)), ["gender", "age"])
         # put group back into evaluation state
         group.status = ContactGroup.STATUS_EVALUATING
@@ -501,27 +497,20 @@ class ContactGroupTest(TembaTest):
     @mock_mailroom
     def test_member_count(self, mr_mocks):
         group = self.create_group("Cool kids")
-
-        # add contacts via the related field
         group.contacts.add(self.joe, self.frank)
 
         self.assertEqual(ContactGroup.user_groups.get(pk=group.pk).get_member_count(), 2)
 
-        # add contacts via update_contacts
-        group.update_contacts(self.user, [self.mary], add=True)
+        group.contacts.add(self.mary)
 
         self.assertEqual(ContactGroup.user_groups.get(pk=group.pk).get_member_count(), 3)
 
-        # remove contacts via update_contacts
-        group.update_contacts(self.user, [self.mary], add=False)
+        group.contacts.remove(self.mary)
 
         self.assertEqual(ContactGroup.user_groups.get(pk=group.pk).get_member_count(), 2)
 
         # blocking a contact removes them from all user groups
         self.joe.block(self.user)
-
-        with self.assertRaises(ValueError):
-            group.update_contacts(self.user, [self.joe], True)
 
         group = ContactGroup.user_groups.get(pk=group.pk)
         self.assertEqual(group.get_member_count(), 1)
@@ -895,14 +884,14 @@ class ContactGroupCRUDLTest(TembaTest):
         ContactGroup.user_groups.filter(id=self.dynamic_group.pk).update(status=ContactGroup.STATUS_READY)
 
         # update both name and query, form should fail, because query is not parsable
-        mr_mocks.parse_query("(!))!)", error="error at !")
+        mr_mocks.error("error at !", code="unexpected_token", extra={"token": "!"})
         response = self.client.post(url, dict(name="Frank", query="(!))!)"))
-        self.assertFormError(response, "form", "query", "error at !")
+        self.assertFormError(response, "form", "query", "Invalid query syntax at '!'")
 
         # try to update a group with an invalid query
-        mr_mocks.parse_query("name <> some_name", error="error at >")
+        mr_mocks.error("error at >", code="unexpected_token", extra={"token": ">"})
         response = self.client.post(url, dict(name="Frank", query="name <> some_name"))
-        self.assertFormError(response, "form", "query", "error at >")
+        self.assertFormError(response, "form", "query", "Invalid query syntax at '>'")
 
         # dependent on id
         response = self.client.post(url, dict(name="Frank", query="id = 123"))
@@ -2896,10 +2885,10 @@ class ContactTest(TembaTest):
             self.assertEqual(history_icon(item), '<span class="glyph icon-bubble-right"></span>')
 
         item = {"type": "flow_entered", "obj": run}
-        self.assertEqual(history_icon(item), '<span class="glyph icon-tree-2"></span>')
+        self.assertEqual(history_icon(item), '<span class="glyph icon-flow"></span>')
 
         run.run_event_type = "Invalid"
-        self.assertEqual(history_icon(item), '<span class="glyph icon-tree-2"></span>')
+        self.assertEqual(history_icon(item), '<span class="glyph icon-flow"></span>')
 
         item = {"type": "flow_exited", "obj": run}
 
@@ -2995,7 +2984,8 @@ class ContactTest(TembaTest):
             self.assertEqual(self.joe, response.context["object"])
             self.assertNotContains(response, "Add Connection")
 
-    def test_read(self):
+    @mock_mailroom
+    def test_read(self, mr_mocks):
         read_url = reverse("contacts.contact_read", args=[self.joe.uuid])
 
         for i in range(5):
@@ -3368,7 +3358,7 @@ class ContactTest(TembaTest):
 
         self.assertEqual(self.joe.user_groups.filter(is_active=True).count(), 2)
 
-        self.client.post(joe_and_frank_filter_url, {"action": "unlabel", "objects": self.joe.id, "add": True})
+        self.client.post(joe_and_frank_filter_url, {"action": "label", "objects": self.joe.id, "add": False})
 
         self.assertEqual(self.joe.user_groups.filter(is_active=True).count(), 2)
 
@@ -3438,10 +3428,10 @@ class ContactTest(TembaTest):
 
         # now let's test removing a contact from a group
         post_data = dict()
-        post_data["action"] = "unlabel"
+        post_data["action"] = "label"
         post_data["label"] = self.joe_and_frank.id
         post_data["objects"] = self.frank.id
-        post_data["add"] = True
+        post_data["add"] = False
         self.client.post(joe_and_frank_filter_url, post_data, follow=True)
         self.assertEqual(len(self.joe_and_frank.contacts.all()), 0)
 
@@ -3571,7 +3561,13 @@ class ContactTest(TembaTest):
         self.joe.save(update_fields=("language",), handle_update=False)
 
         # add some languages to our org, but not french
-        self.client.post(reverse("orgs.org_languages"), dict(primary_lang="hat", languages="arc,spa"))
+        self.client.post(
+            reverse("orgs.org_languages"),
+            dict(
+                primary_lang='{"name":"Haitian", "value":"hat"}',
+                languages=['{"name":"Official Aramaic", "value":"arc"}', '{"name":"Spanish", "value":"spa"}'],
+            ),
+        )
 
         response = self.client.get(reverse("contacts.contact_update", args=[self.joe.id]))
         self.assertContains(response, "French (Missing)")
@@ -4931,14 +4927,15 @@ class ContactTest(TembaTest):
         self.login(self.admin)
         self.create_campaign()
 
-        ballers = self.create_group("Ballers")
+        contact = self.create_contact("Joe", urn="tel:123")
+        ballers = self.create_group("Ballers", contacts=[contact])
 
         self.campaign.group = ballers
         self.campaign.save()
 
         field_created_on = self.org.contactfields.get(key="created_on")
 
-        self.created_on_event = CampaignEvent.create_flow_event(
+        created_on_event = CampaignEvent.create_flow_event(
             self.org,
             self.admin,
             self.campaign,
@@ -4948,22 +4945,10 @@ class ContactTest(TembaTest):
             flow=self.reminder_flow,
         )
 
-        event_fires = EventFire.objects.filter(event=self.created_on_event)
-        self.assertEqual(event_fires.count(), 0)
+        EventFire.create_eventfires_for_event(created_on_event)
 
-        contact, urn = Contact.get_or_create(self.org, "tel:123", user=self.user, name="Joe")
-
-        event_fires = EventFire.objects.filter(event=self.created_on_event)
-        self.assertEqual(event_fires.count(), 0)
-
-        ballers.update_contacts(self.admin, [contact], add=True)
-
-        event_fires = EventFire.objects.filter(event=self.created_on_event)
+        event_fires = EventFire.objects.filter(event=created_on_event)
         self.assertEqual(event_fires.count(), 1)
-
-        event_fire = EventFire.objects.filter(event=self.created_on_event).first()
-        contact_created_on = contact.created_on.replace(second=0, microsecond=0)
-        self.assertEqual(event_fire.scheduled, contact_created_on + timedelta(minutes=5))
 
     def test_contact_import_handle_update_contact(self):
         self.login(self.admin)